CHORE: Reorganizing the file directory structure
