--- examples/ReactNativeSpacingSystemExample/src/Stack.tsx
@@ -1,12 +1,0 @@
-import * as React from "react";
-import { Stack as LibStack, StackProps } from "react-native-spacing-system";
-import { SpacingKeys, spacing } from "./spacing";
-
-export type Props = Omit<StackProps, "size"> & { size: SpacingKeys };
-
-const Stack = (props: Props): React.FunctionComponentElement<StackProps> => {
-  const { size, ...debugProps } = props;
-  return <LibStack size={spacing[size]} {...debugProps} />;
-};
-
-export default Stack;
