+++ examples/ReactNativeSpacingSystemExample/src/constants/colors.ts
