+++ examples/ReactNativeSpacingSystemExample/src/constants/fontSize.ts
